fix: added enum in domain, coloredStatus component used and seperator refactored.
