+++ lib/public/views/Environments/format/displayEnvironmentStatusHistory.js
@@ -11,6 +11,8 @@
  *  or submit itself to any jurisdiction.
  */
 import { h } from '/js/src/index.js';
+import { STATUS_ACRONYMS } from '../../../domain/enums/statusAcronyms.js';
+import { coloredEnvironmentStatusComponent } from '../ColoredEnvironmentStatusComponent.js';
 
 /**
  * Display the given environment's status history string
@@ -21,48 +23,15 @@ import { h } from '/js/src/index.js';
 export const displayEnvironmentStatusHistory = (environment) => {
     const { historyItems } = environment;
 
-    const statusAcronyms = {
-        STANDBY: 'S',
-        DEPLOYED: 'DL',
-        CONFIGURED: 'C',
-        RUNNING: 'R',
-        ERROR: 'E',
-        MIX: 'M',
-        DESTROYED: 'DS',
-    };
-
-    /**
-     * Get correct color class from environmentname.
-     * @param {*} status One of the Environment statuses
-     * @returns {string} The classname used to color the acronoym.
-     */
-    const getColorClassEnvironment = (status) => {
-        switch (status) {
-            case 'RUNNING':
-                return '.success';
-            case 'ERROR':
-                return '.danger';
-            case 'CONFIGURED':
-                return '.warning';
-            default:
-                return '';
-        }
-    };
-
-    const statusHistory = [];
-
-    // Iterate through historyItems
-    for (const { status } of historyItems) {
-        if (status in statusAcronyms) {
-            statusHistory.push({ status, content: statusAcronyms[status] });
-            statusHistory.push({ status: null, content: '-' });
-        }
-    }
-
-    statusHistory.pop();
+    const statusHistory = historyItems
+        .filter(({ status }) => status in STATUS_ACRONYMS)
+        .map(({ status }) => ({ status, content: STATUS_ACRONYMS[status] }));
 
     return h(
         '.flex-row',
-        statusHistory.map((value) => h(getColorClassEnvironment(value.status), value.content)),
+        statusHistory.map((value, index) => h(".flex-row", [
+            coloredEnvironmentStatusComponent(value.status, value.content),
+            h("", ((index < statusHistory.length - 1) ? "-" : "")),
+        ])),
     );
 };
